requirements.txt
joblib==1.2.0
numpy==1.23.4
pandas==1.5.0
python-dateutil==2.8.2
pytz==2022.4
scikit-learn==1.1.2
scipy==1.9.2
six==1.16.0
threadpoolctl==3.1.0
psycopg2-binary==2.9.10
fastapi==0.115.6
pymongo==4.10.1
python-multipart==0.0.20
uvicorn==0.34.0
setuptools==58.1.0
requests==2.32.3
pytest==8.3.4
httpx==0.28.1